Decomposed TaskDetail component
